globals: type per-TU sprite-handle globals as struct Sprite* (byte-identical)
Re-type 67 single-TU sprite-handle globals (unsigned int/void* -> struct Sprite*)
across balloonz, carousel, castle, castle_level, draw, driving_school, entrance,
joust, jungle_cruise, plane_ride, safari_ride, shops, temple, temple_slide,
tilemap, water_works, western_town, space_tower, spinning_barrels, text.
water_works FUN_004186b0 returns (unsigned int)DAT_004cc020 (uint ABI).
Each global verified used only as a sprite handle; aggregate holds 53.02%.

+++ src/legoland/water_works.c
@@ -376,7 +376,7 @@ unsigned int FUN_004186b0(struct WaterArg *arg) {
     DAT_004cbfec = (unsigned int)DAT_004cbfdc->field_64;
     // STRING: LEGOLAND 0x004b5078
     DAT_004cc020 = LoadSprite("ww_elsquirt.lls", 1);
-    return DAT_004cc020;
+    return (unsigned int)DAT_004cc020;
 }
 
 // FUNCTION: LEGOLAND 0x004186f0
